+++ controllers/index.js
@@ -1,4 +1,5 @@
 var gdax = require('../lib/gdax');
+var get = require('lodash').get;
 
 const exchange = gdax.getInstance();
 
@@ -31,6 +32,26 @@ module.exports.getOrders = (req, res, next) => {
 
 module.exports.getOrder = (req, res, next) => {
   exchange.authedClient.getOrder(req.params.id, (err, resp, data) => {
+    if (err) {
+      return res.status(500).json({error: err});
+    }
+    return res.status(200).json(data);
+  });
+}
+
+module.exports.cancelOrders = (req, res, next) => {
+  let product = get(req.body, 'product_id', null);
+  let options = product ? {product_id: product} : {};
+  exchange.authedClient.cancelAllOrders(options, (err, resp, data) => {
+    if (err) {
+      return res.status(500).json({error: err});
+    }
+    return res.status(200).json(data);
+  });
+}
+
+module.exports.cancelOrder = (req, res, next) => {
+  exchange.authedClient.cancelOrder(req.params.id, (err, resp, data) => {
     if (err) {
       return res.status(500).json({error: err});
     }